rights

+++ backend/src/main/java/com/example/backend/trida/TridaController.java
@@ -1,9 +1,9 @@
 package com.example.backend.trida;
 
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,6 +12,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.example.backend.security.Helper;
+import com.example.backend.security.UserToken;
+import com.example.backend.security.user.UserEntity;
+import com.example.backend.security.user.UserRepository;
 
 @RestController
 @RequestMapping("/tridy")
@@ -22,10 +27,16 @@ public class TridaController {
 
     @GetMapping("/{id}")
     public @ResponseBody ResponseEntity<TridaEntity> get(@PathVariable long id) {
-        var t = tridaRepository.findById(id);
-        if (t.isPresent()) {
-            return new ResponseEntity<>(t.get(), HttpStatus.OK);
-        } else {
+        try {
+            var trida = tridaRepository.findById(id).get();
+            if (Helper.hasRightsForTrida(trida)) {
+                return new ResponseEntity<>(trida, HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+
+        } catch (Exception e) {
+            System.out.println(e);
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
@@ -35,30 +46,56 @@ public class TridaController {
         var trida = new TridaEntity();
         trida.setName(name);
 
+        trida.setUserEntity(Helper.getUserEntity());
+
         tridaRepository.save(trida);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PostMapping("/delete")
     public @ResponseBody ResponseEntity<String> delete(@RequestParam long id) {
-        tridaRepository.deleteById(id);
+        try {
+            var tridaToDelete = tridaRepository.findById(id).get();
 
-        return new ResponseEntity<>(HttpStatus.OK);
+            if (Helper.hasRightsForTrida(tridaToDelete)) {
+                tridaRepository.deleteById(id);
+                return new ResponseEntity<>(HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @GetMapping("/all")
     public @ResponseBody ResponseEntity<Iterable<TridaEntity>> getAll() {
-        return new ResponseEntity<>(tridaRepository.findAll(), HttpStatus.OK);
+
+        return new ResponseEntity<>(tridaRepository.findAllByUserEntity(Helper.getUserEntity()), HttpStatus.OK);
     }
 
     @PostMapping("/update")
     public @ResponseBody ResponseEntity<String> update(@RequestParam long id,
             @RequestParam(required = false) String name) {
-        var trida = tridaRepository.findById(id).get();
-        trida.setName(name);
-        tridaRepository.save(trida);
+        try {
+            var trida = tridaRepository.findById(id).get();
 
-        return new ResponseEntity<>(HttpStatus.OK);
+            if (Helper.hasRightsForTrida(trida)) {
+                trida.setName(name);
+                tridaRepository.save(trida);
+
+                return new ResponseEntity<>(HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
     }
+
+    
 
 }
